Fixed V4 format design bug
Problem was that we needed a valid passphrase to
detect the end of and integrity of the keyblocks.
This meant that an incorrect passphrase would cause
keyblocks to be parsed where there weren't any, resulting
in crazy nIter values... Not Good.

Solution was to add an 'end of keyblock' marker that
wasn't dependent on the keyblocks. Since plaintext was
frowned upon in V3, we've done things slightly more interesting.
See Section 2.3 of the format doc for detatils.