Fix BR1220 - crash on Fedora 21 - Linux ONLY
The root cause is that __FILE__ and __FUNCTION__ are "C string constants",
which are char *, not wchar_t, yet we were formatting this as %S.

This commit fixes the issue locally, but:
(a) This needs to be tested on other Linux platforms
(b) Windows broken, hence in separate branch.
(c) Getting rid of all the #ifdef UNICODE stuff is the Right Way to
    fix this long-term.

Also removed the unused _FMT macro, and tweaked the initial guess value
in GetStringBufSize to something more reasonable.